Refactor(jisu/notification): 알림 테스트 파일에서 게시판 관련 알림 분리 #117
